requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
llama-index==0.9.48
openai>=1.1.0
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.6.1
setuptools>=65.5.1
pypdf==3.17.1
sentence-transformers==2.2.2
transformers==4.36.2
httpx==0.23.3
chromadb==0.4.22 